Clean up stray text in test file
