+++ src/components/comments/CommentCard.vue
@@ -205,9 +205,8 @@
 </template>
 
 <script setup>
-import { ref, computed } from 'vue'
+
 import {
-  UserIcon,
   SparklesIcon,
   TrashIcon,
   ClipboardIcon,
@@ -218,28 +217,11 @@ import {
   CheckBadgeIcon,
   ExclamationTriangleIcon,
   CpuChipIcon,
-  ClockIcon,
-  // ✅ Nuevos iconos para estados emocionales
-  FaceSmileIcon,
-  BoltIcon,
-  CloudIcon,
-  EyeIcon,
-  HandRaisedIcon,
-  BeakerIcon
+  ClockIcon
 } from '@heroicons/vue/24/outline'
-import {
-  SparklesIcon as SparklesIconSolid,
-  UserIcon as UserIconSolid,
-  // ✅ Iconos sólidos para estados emocionales
-  FaceSmileIcon as FaceSmileIconSolid,
-  BoltIcon as BoltIconSolid,
-  CloudIcon as CloudIconSolid,
-  EyeIcon as EyeIconSolid,
-  HandRaisedIcon as HandRaisedIconSolid,
-  BeakerIcon as BeakerIconSolid,
-  HeartIcon as HeartIconSolid
-} from '@heroicons/vue/24/solid'
+import { UserIcon as UserIconSolid } from '@heroicons/vue/24/solid'
 import ChibiAvatar from '../ui/ChibiAvatar.vue'
+import { useCommentCard } from '@/composables/useCommentCard'
 
 // Props
 const props = defineProps({
@@ -262,161 +244,23 @@ const props = defineProps({
 
 // Emits
 const emit = defineEmits(['delete', 'ask-ai', 'update-feedback'])
-
-// State
-const copied = ref(false)
-const showDeleteConfirm = ref(false)
-const deleting = ref(false)
-
-// Computed
-const commentTypeClass = computed(() => {
-  return props.comment.role === 'assistant'
-    ? 'bg-gradient-to-br from-purple-50 to-white border-l-4 border-l-purple-500 group hover:shadow-lg hover:-translate-y-0.5'
-    : 'bg-gradient-to-br from-blue-50 to-white border-l-4 border-l-blue-500 group hover:shadow-lg hover:-translate-y-0.5'
-})
-
-const avatarBgClass = computed(() => {
-  return props.comment.role === 'assistant'
-    ? 'bg-gradient-to-br from-purple-100 to-purple-200'
-    : 'bg-gradient-to-br from-blue-100 to-blue-200'
-})
-
-const authorClass = computed(() => {
-  return props.comment.role === 'assistant'
-    ? 'text-purple-800'
-    : 'text-blue-800'
-})
-
-// Methods
-const formatRelativeTime = (dateString) => {
-  const now = new Date()
-  const commentDate = new Date(dateString)
-  const diffInMinutes = Math.floor((now - commentDate) / (1000 * 60))
-
-  if (diffInMinutes < 1) return 'ahora'
-  if (diffInMinutes < 60) return `hace ${diffInMinutes}m`
-
-  const diffInHours = Math.floor(diffInMinutes / 60)
-  if (diffInHours < 24) return `hace ${diffInHours}h`
-
-  const diffInDays = Math.floor(diffInHours / 24)
-  if (diffInDays < 7) return `hace ${diffInDays}d`
-
-  return commentDate.toLocaleDateString('es-ES')
-}
-
-const copyContent = async () => {
-  try {
-    await navigator.clipboard.writeText(props.comment.message)
-    copied.value = true
-    setTimeout(() => {
-      copied.value = false
-    }, 2000)
-  } catch (err) {
-    console.error('Error copiando contenido:', err)
-  }
-}
-
-const deleteComment = () => {
-  showDeleteConfirm.value = true
-}
-
-const confirmDelete = async () => {
-  deleting.value = true
-  try {
-    emit('delete', props.comment.id)
-    showDeleteConfirm.value = false
-  } catch (err) {
-    console.error('Error eliminando comentario:', err)
-  } finally {
-    deleting.value = false
-  }
-}
-
-const askAI = () => {
-  console.log('💬 CommentCard - Solicitando consejo de AI para:', props.comment.message)
-  emit('ask-ai', props.comment.message)
-}
-
-// ✅ Corrección del bug en la función toggleReaction
-const toggleReaction = (type) => {
-  const feedbackData = {}
-
-  switch (type) {
-    case 'useful':
-      if (props.comment.role === 'user') {
-        feedbackData.user_is_useful = !props.comment.user_is_useful
-      } else {
-        feedbackData.assistant_is_useful = !props.comment.assistant_is_useful
-      }
-      break
-    case 'precise':
-      feedbackData.assistant_is_precise = !props.comment.assistant_is_precise
-      break
-    case 'grateful':
-      if (props.comment.role === 'user') {
-        feedbackData.user_is_grateful = !props.comment.user_is_grateful // ✅ CORREGIDO
-      } else {
-        feedbackData.assistant_is_grateful = !props.comment.assistant_is_grateful // ✅ CORREGIDO
-      }
-      break
-  }
-
-  console.log('👍 CommentCard - Actualizando feedback:', {
-    commentId: props.comment.id,
-    type,
-    feedbackData
-  })
-
-  emit('update-feedback', props.comment.id, feedbackData)
-}
-
-// ✅ NUEVA función con iconos en lugar de emojis
-const getEmotionalStateIcon = (state) => {
-  const icons = {
-    happy: FaceSmileIconSolid,
-    excited: SparklesIconSolid,
-    calm: CloudIconSolid,
-    focused: EyeIconSolid,
-    supportive: HandRaisedIconSolid,
-    encouraging: HeartIconSolid,
-    thoughtful: BeakerIconSolid,
-    energetic: BoltIconSolid
-  }
-  return icons[state] || FaceSmileIconSolid
-}
-
-// ✅ NUEVA función para colores de estados emocionales
-const getEmotionalStateColor = (state) => {
-  const colors = {
-    happy: 'text-yellow-500',
-    excited: 'text-orange-500',
-    calm: 'text-blue-500',
-    focused: 'text-purple-500',
-    supportive: 'text-green-500',
-    encouraging: 'text-pink-500',
-    thoughtful: 'text-indigo-500',
-    energetic: 'text-red-500'
-  }
-  return colors[state] || 'text-gray-500'
-}
-
-// ✅ NUEVA función para formatear nombres de modelos
-const formatModelName = (modelName) => {
-  // Mapeo de nombres de modelos para mostrar versiones más amigables
-  const modelDisplayNames = {
-    'gpt-4': 'GPT-4',
-    'gpt-4-turbo': 'GPT-4 Turbo',
-    'gpt-4o': 'GPT-4o',
-    'gpt-4o-mini': 'GPT-4o Mini',
-    'gpt-3.5-turbo': 'GPT-3.5',
-    'claude-3-sonnet': 'Claude 3',
-    'claude-3-haiku': 'Claude 3 Haiku',
-    'gemini-pro': 'Gemini Pro'
-  }
-
-  return modelDisplayNames[modelName] || modelName
-}
+const {
+  copied,
+  showDeleteConfirm,
+  deleting,
+  commentTypeClass,
+  avatarBgClass,
+  authorClass,
+  formatRelativeTime,
+  copyContent,
+  deleteComment,
+  confirmDelete,
+  askAI,
+  toggleReaction,
+  getEmotionalStateIcon,
+  getEmotionalStateColor,
+  formatModelName
+} = useCommentCard(props, emit)
 </script>
 
 <style scoped>
